requirements.txt
websockets==2.4
pyserial==2.7
